refactor proxy settings
