+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_core/_events.py
@@ -20,10 +20,8 @@ from servicelib.json_serialization import json_dumps
 from servicelib.rabbitmq import RabbitMQClient
 from simcore_postgres_database.models.comp_tasks import NodeClass
 
-from .....core.dynamic_sidecar_settings import (
-    DynamicSidecarProxySettings,
-    DynamicSidecarSettings,
-)
+from .....core.dynamic_services.proxy import DynamicSidecarProxySettings
+from .....core.dynamic_sidecar_settings import DynamicSidecarSettings
 from .....models.dynamic_services_scheduler import (
     DockerContainerInspect,
     DockerStatus,
